Handle `on_url` being called multiple times on the same request
In some cases -- specifically when there's TCP packet fragmentation -- 
the `on_url` callback in httptools can be called multiple times in a
single request. In that case the URL has to be built from the separate
calls.

We now use the `on_message_begin`-callback to reset the per-request
state and do the URL parsing in `on_headers_complete` where we
know that all of the URL has been received.